Move status controller into api namespace.
This will make the status endpoint accessible via this URL:

http://ohana-api-demo.herokuapp.com/api/.well-known/status

If you are using a subdomain with a custom domain, it would look
something like this:

http://api.smc-connect.org/.well-known/status
